style: restructure project files outside of app folder to make routable files senses
