+++ adapters/treeAdapter.ts
