+++ app/api/graphql/route.ts
@@ -2,8 +2,8 @@ import { ApolloServer } from '@apollo/server';
 import { startServerAndCreateNextHandler } from '@as-integrations/next';
 import { NextRequest } from 'next/server';
 
-import typeDefs from '../../graphql/typeDefs';
-import { resolvers } from '../../graphql/resolvers';
+import typeDefs from '../../../graphql/typeDefs';
+import { resolvers } from '../../../graphql/resolvers';
 
 const apolloServer = new ApolloServer({ typeDefs, resolvers });
